moved pin connection tracking into class, moved tracking from NodeGraph into its viewmodel
